# perf(crypto/merkle): reduce allocations in HashFromByteSlices