@cristaloleg 2024-05-25
I see 3 approvals but also 2 failed jobs, can someone restart them? (I don't have access neither to CI nor to merge button). Thanks in advance 😄